v1.3.0—add a simple tracker process implementation.
Sometimes, you don't want to litter your model with SUT-data. In this case
you can employ the new tracker process code to push the SUT-data into
a separate process which then tracks the state for the SUT. It is often
simpler to do this than it is to write ones own process from scratch.